backend/requirements.txt
# FastAPI Core
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# LangChain & RAG
langchain>=0.1.0
langchain-anthropic>=0.1.0
langchain-community>=0.0.10

# Vector Database & Documents
chromadb>=0.4.0
pypdf>=3.17.0

# Local Embeddings (HuggingFace)
sentence-transformers>=2.2.2
transformers>=4.35.0
torch>=2.1.0

# Utilities
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
